forsrc-common

+++ src/main/java/com/forsrc/common/tool/ToolDateTime.java
@@ -43,72 +43,72 @@ public class ToolDateTime {
 
   // <<<----------------------- plus LocalDateTime -----------------------
 
-  public static Date plusSeconds(LocalDateTime localDateTime, int seconds) {
+  public static Date plusSeconds(LocalDateTime localDateTime, long seconds) {
     LocalDateTime newDatetime = localDateTime.plusSeconds(seconds);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusSeconds(LocalDateTime localDateTime, int seconds) {
+  public static Date minusSeconds(LocalDateTime localDateTime, long seconds) {
     LocalDateTime newDatetime = localDateTime.minusSeconds(seconds);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusMinutes(LocalDateTime localDateTime, int minutes) {
+  public static Date plusMinutes(LocalDateTime localDateTime, long minutes) {
     LocalDateTime newDatetime = localDateTime.plusMinutes(minutes);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusMinutes(LocalDateTime localDateTime, int minutes) {
+  public static Date minusMinutes(LocalDateTime localDateTime, long minutes) {
     LocalDateTime newDatetime = localDateTime.minusMinutes(minutes);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusHours(LocalDateTime localDateTime, int hours) {
+  public static Date plusHours(LocalDateTime localDateTime, long hours) {
     LocalDateTime newDatetime = localDateTime.plusHours(hours);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusHours(LocalDateTime localDateTime, int hours) {
+  public static Date minusHours(LocalDateTime localDateTime, long hours) {
     LocalDateTime newDatetime = localDateTime.minusHours(hours);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusDays(LocalDateTime localDateTime, int days) {
+  public static Date plusDays(LocalDateTime localDateTime, long days) {
     LocalDateTime newDatetime = localDateTime.plusDays(days);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusDays(LocalDateTime localDateTime, int days) {
+  public static Date minusDays(LocalDateTime localDateTime, long days) {
     LocalDateTime newDatetime = localDateTime.minusDays(days);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusWeeks(LocalDateTime localDateTime, int weeks) {
+  public static Date plusWeeks(LocalDateTime localDateTime, long weeks) {
     LocalDateTime newDatetime = localDateTime.plusWeeks(weeks);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusWeeks(LocalDateTime localDateTime, int weeks) {
+  public static Date minusWeeks(LocalDateTime localDateTime, long weeks) {
     LocalDateTime newDatetime = localDateTime.minusWeeks(weeks);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusMonths(LocalDateTime localDateTime, int months) {
+  public static Date plusMonths(LocalDateTime localDateTime, long months) {
     LocalDateTime newDatetime = localDateTime.plusMonths(months);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusMonths(LocalDateTime localDateTime, int months) {
+  public static Date minusMonths(LocalDateTime localDateTime, long months) {
     LocalDateTime newDatetime = localDateTime.minusMonths(months);
     return toDateTime(newDatetime);
   }
 
-  public static Date plusYears(LocalDateTime localDateTime, int years) {
+  public static Date plusYears(LocalDateTime localDateTime, long years) {
     LocalDateTime newDatetime = localDateTime.plusYears(years);
     return toDateTime(newDatetime);
   }
 
-  public static Date minusYears(LocalDateTime localDateTime, int years) {
+  public static Date minusYears(LocalDateTime localDateTime, long years) {
     LocalDateTime newDatetime = localDateTime.minusYears(years);
     return toDateTime(newDatetime);
   }
@@ -117,42 +117,42 @@ public class ToolDateTime {
 
   // <<<----------------------- plus LocalDate -----------------------
 
-  public static Date plusDays(LocalDate localDate, int days) {
+  public static Date plusDays(LocalDate localDate, long days) {
     LocalDate newDate = localDate.plusDays(days);
     return toDateTime(newDate);
   }
 
-  public static Date minusDays(LocalDate localDate, int days) {
+  public static Date minusDays(LocalDate localDate, long days) {
     LocalDate newDate = localDate.minusDays(days);
     return toDateTime(newDate);
   }
 
-  public static Date plusWeeks(LocalDate localDate, int weeks) {
+  public static Date plusWeeks(LocalDate localDate, long weeks) {
     LocalDate newDate = localDate.plusWeeks(weeks);
     return toDateTime(newDate);
   }
 
-  public static Date minusWeeks(LocalDate localDate, int weeks) {
+  public static Date minusWeeks(LocalDate localDate, long weeks) {
     LocalDate newDate = localDate.minusWeeks(weeks);
     return toDateTime(newDate);
   }
 
-  public static Date plusMonths(LocalDate localDate, int months) {
+  public static Date plusMonths(LocalDate localDate, long months) {
     LocalDate newDate = localDate.plusMonths(months);
     return toDateTime(newDate);
   }
 
-  public static Date minusMonths(LocalDate localDate, int months) {
+  public static Date minusMonths(LocalDate localDate, long months) {
     LocalDate newDate = localDate.minusMonths(months);
     return toDateTime(newDate);
   }
 
-  public static Date plusYears(LocalDate localDate, int years) {
+  public static Date plusYears(LocalDate localDate, long years) {
     LocalDate newDate = localDate.plusYears(years);
     return toDateTime(newDate);
   }
 
-  public static Date minusYears(LocalDate localDate, int years) {
+  public static Date minusYears(LocalDate localDate, long years) {
     LocalDate newDate = localDate.minusYears(years);
     return toDateTime(newDate);
   }
